+++ src/osipparser2/osip_message_to_str.c
@@ -972,9 +972,9 @@ _osip_message_to_str (osip_message_t * sip, char **dest, size_t *message_length,
       size_t size = message - *dest;
       int offset_of_body;
       int offset_content_length_to_modify = 0;
-      offset_of_body = start_of_bodies - *dest;
+      offset_of_body = (int)(start_of_bodies - *dest);
       if (content_length_to_modify != NULL)
-	offset_content_length_to_modify = content_length_to_modify - *dest;
+	offset_content_length_to_modify = (int)(content_length_to_modify - *dest);
       malloc_size = message - *dest + body_length + 100;
       *dest = osip_realloc (*dest, malloc_size);
       if (*dest == NULL)
